@@ -1,33 +1,41 @@
 [project]
 name = "mcp-server-telegram"
 version = "0.1.0"
-description = "MCP Server providing a tool to post messages to a Telegram channel."
-
+description = "MCP Server for sending messages to Telegram channels"
 authors = [
-    { name = "Your Name", email = "your.email@example.com" },
+    { name = "XyAgent", email = "agent@xyber.inc" },
 ]
 readme = "README.md"
-requires-python = ">=3.11"
-
+requires-python = ">=3.12"
 
 dependencies = [
-    "mcp>=1.6.0,<2.0.0",
-    "starlette>=0.37.0",
-    "pydantic-settings>=2.0",
-    "python-dotenv",
+    "mcp>=1.8.0,<2.0.0",
+    "pydantic-settings>=2.0",  
     "uvicorn[standard]",
-    "requests",  # Added for Telegram API calls
-    "langchain_mcp_adapters",  
+    "fastmcp>=2.3.4",
+    "fastapi>=0.115.12",
+    # Telegram-specific dependencies
+    "python-telegram-bot>=20.0",
 ]
 
 [dependency-groups]
 dev = [
     "pytest",
-    "httpx",
+    "httpx", 
     "ruff",
     "black",
     "isort",
 ]
+debug = [
+    "debugpy>=1.8.0",
+]
+
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
+[tool.hatch.build.targets.wheel]
+packages = ["src/mcp_server_telegram"]
 
 # --- Tool Configurations --- #
 
